+++ ui/package.json
@@ -2,7 +2,7 @@
   "name": "webpack_example_2",
   "version": "1.0.0",
   "description": "",
-  "main": "app.js",
+  "main": "index.js",
   "scripts": {
     "start": "webpack-dev-server",
     "build": "webpack",
@@ -20,7 +20,7 @@
     "bootstrap": "^3.3.7",
     "css-loader": "^0.28.7",
     "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "react-dom": "^15.6.2",
     "react-redux": "^5.0.6",
     "react-router": "^4.2.0",
     "react-router-redux": "^4.0.8",
