+++ ui/public/js/a_dream.js
@@ -1,5 +1,8 @@
-const React = require('react');
+import React from 'react'
+import ReactDOM from 'react-dom';
+import { Router, IndexRoute, Route, hashHistory } from 'react-router'
 const storiesOf = require('@storybook/react').storiesOf;
+
 import 'bootstrap/dist/css/bootstrap.css';
 import '../css/app.css'
