+++ ui/public/js/app.js
@@ -1,34 +1,23 @@
-const React = require('react');
-const ReactDOM = require('react-dom');
+import React from 'react'
+import ReactDOM from 'react-dom';
+import { Router, IndexRoute, Route, hashHistory } from 'react-router'
 
-const DreamForm = require('./add_dream')
-const DreamFeed = require('./dream_feed')
-const Dream = require('./a_dream')
-const NavBar = require('./header')
-const Footer = require('./footer')
+import NavBar from './header'
 
 import 'bootstrap/dist/css/bootstrap.css';
 import '../css/app.css'
-
-//App
 
 const App = React.createClass({
   render: function () {
     return (
       <div>
+        <div className="container nav_contain">
+        <NavBar />
+        </div>
 
-      <div className="container nav_contain">
-      <NavBar />
-      </div>
-
-      <div className="container app">
-      <DreamForm charCount={ 140 } />
-      <DreamFeed />
-      </div>
-
-      <div className="container foot_contain">
-      <Footer />
-      </div>
+        <div className="row">
+            {this.props.children}
+        </div>
 
       </div>
 
@@ -36,12 +25,4 @@ const App = React.createClass({
   }
 })
 
-// Wait for the window to load
-window.onload = function () {
-  const DreamApp = document.querySelector('#App');
-
-  ReactDOM.render(
-    <App/>,
-    DreamApp
-  );
-}
+module.exports = App;
