+++ ui/public/js/dream_feed.js
@@ -1,6 +1,10 @@
-const React = require('react');
+import React from 'react'
+import ReactDOM from 'react-dom';
+import { Router, IndexRoute, Route, hashHistory } from 'react-router'
 const storiesOf = require('@storybook/react').storiesOf;
-const Dream = require('./a_dream')
+
+import Dream from './a_dream'
+
 import 'bootstrap/dist/css/bootstrap.css';
 import '../css/app.css'
